feat: enhance role management with improved delete confirmation and user feedback

+++ app/Livewire/Konfigurasi/Aksesibilitas/Roles/Detail/DaftarRolesSub.php
@@ -4,6 +4,7 @@ namespace App\Livewire\Konfigurasi\Aksesibilitas\Roles\Detail;
 
 use App\Models\User;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Attributes\On;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -21,9 +22,24 @@ class DaftarRolesSub extends Component
     ];
 
     #[On('aksesibilitas.roles.detail.delete')]
-    public function deleteRole($user_id)
+    public function deleteRole($id, $flag)
     {
-        $user = User::find($user_id);
+        $user = User::find($id);
+
+        if (!$user) {
+            $this->dispatch('error', 'User tidak ditemukan.');
+            return;
+        }
+        if ($user->id === Auth::id()) {
+            $this->dispatch('error', 'Tidak bisa menghapus Akun saat ini.');
+            return;
+        }
+
+        if ($flag === 'confirm') {
+            $this->dispatch('swal-confirm', $id, 'aksesibilitas.roles.detail.delete', ['text' => "Data User roles {$user->name} yang dihapus tidak dapat dikembalikan"]);
+            return;
+        }
+
         if ($user) {
             $user->roles()->detach();
             $this->dispatch('swal', 'Role has been deleted successfully');
